chore(): travis update
