+++ .travis.yml
@@ -2,13 +2,18 @@ language: node_js
 node_js:
   - 9
   - 6
-env:
-  - SECRET_KEY=#(@#GYzU#EWERGFXTSWEW(@#YW($EW(
-  global:
-    - export NODE_ENV=test
-  script:
-  - npm install
+cache:
+  directories:
+  - node_modules
+addons:
+- postgresql: "9.4"
+services:
+- postgresql
+global:
+  - export NODE_ENV=test
+before_script:
+- npm install
+script:
   - npm run test
   - npm run coverage
-
-  after_success: 'npm run coveralls'
+after_success: 'npm run coveralls'
