+++ client/public/js/hot/hot-update.js
@@ -1,11 +1,1 @@
-webpackHotUpdate(0,{
-
-/***/ 1657:
-/***/ (function(module, exports, __webpack_require__) {
-
-"use strict";
-eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(14);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(15);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(16);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _propTypes = __webpack_require__(4);\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _redux = __webpack_require__(23);\n\nvar _shortid = __webpack_require__(37);\n\nvar _shortid2 = _interopRequireDefault(_shortid);\n\nvar _reactDocumentTitle = __webpack_require__(68);\n\nvar _reactDocumentTitle2 = _interopRequireDefault(_reactDocumentTitle);\n\nvar _Dialog = __webpack_require__(82);\n\nvar _Dialog2 = _interopRequireDefault(_Dialog);\n\nvar _RaisedButton = __webpack_require__(174);\n\nvar _RaisedButton2 = _interopRequireDefault(_RaisedButton);\n\nvar _FlatButton = __webpack_require__(85);\n\nvar _FlatButton2 = _interopRequireDefault(_FlatButton);\n\nvar _modeEdit = __webpack_require__(281);\n\nvar _modeEdit2 = _interopRequireDefault(_modeEdit);\n\nvar _delete = __webpack_require__(280);\n\nvar _delete2 = _interopRequireDefault(_delete);\n\nvar _loader = __webpack_require__(113);\n\nvar _activeCenterAction = __webpack_require__(1658);\n\nvar _deleteCenterAction = __webpack_require__(1659);\n\nvar _currentEventForCenter = __webpack_require__(1660);\n\nvar _currentEventForCenter2 = _interopRequireDefault(_currentEventForCenter);\n\nvar _RecommCenter = __webpack_require__(1661);\n\nvar _RecommCenter2 = _interopRequireDefault(_RecommCenter);\n\nvar _EventModal = __webpack_require__(1662);\n\nvar _EventModal2 = _interopRequireDefault(_EventModal);\n\nvar _editCenterForm = __webpack_require__(1663);\n\nvar _editCenterForm2 = _interopRequireDefault(_editCenterForm);\n\nvar _fetchCenterRelatedTo = __webpack_require__(1664);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  CenterDetail: {\n    displayName: 'CenterDetail'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/centerDetail.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/centerDetail.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar CenterDetail = _wrapComponent('CenterDetail')(function (_Component) {\n  _inherits(CenterDetail, _Component);\n\n  function CenterDetail(props) {\n    _classCallCheck(this, CenterDetail);\n\n    var _this = _possibleConstructorReturn(this, (CenterDetail.__proto__ || Object.getPrototypeOf(CenterDetail)).call(this, props));\n\n    _this.state = {\n      isLoading: true,\n      openAlert: false,\n      open: false,\n      openCenterEvents: false,\n      activeCenter: {\n        centr: {\n          title: 'center'\n        }\n      }\n    };\n\n    _this.handleOpen = _this.handleOpen.bind(_this);\n    _this.handleClose = _this.handleClose.bind(_this);\n    _this.handleAlertOpen = _this.handleAlertOpen.bind(_this);\n    _this.handleAlertClose = _this.handleAlertClose.bind(_this);\n    _this.handleEventsOpen = _this.handleEventsOpen.bind(_this);\n    _this.handleEventsClose = _this.handleEventsClose.bind(_this);\n    return _this;\n  }\n\n  _createClass(CenterDetail, [{\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      $('.modal').modal('close');\n      $('.tooltipped').tooltip({ delay: 50 });\n      var params = this.props.params;\n\n      this.props.fetchCenterAction(params.id);\n    }\n  }, {\n    key: 'componentWillReceiveProps',\n    value: function componentWillReceiveProps(newProps) {\n      if (this.props.params.id !== newProps.params.id) {\n        newProps.fetchCenterAction(newProps.params.id);\n      }\n\n      if (typeof newProps.activeCenterDetail.center !== 'undefined') {\n        newProps.activeCenterDetail.center.events = newProps.activeCenterDetail.events;\n        if (newProps.activeCenterDetail.center.events) {\n          delete newProps.activeCenterDetail.events;\n        }\n        this.setState({ isLoading: false, activeCenter: newProps.activeCenterDetail });\n      }\n    }\n  }, {\n    key: 'editCenter',\n    value: function editCenter() {\n      this.props.editCenterRequestAction();\n    }\n  }, {\n    key: 'handleOpen',\n    value: function handleOpen() {\n      this.props.editCenterRequestAction();\n      this.setState({ open: true });\n    }\n  }, {\n    key: 'handleClose',\n    value: function handleClose() {\n      this.setState({ open: false });\n    }\n  }, {\n    key: 'handleAlertOpen',\n    value: function handleAlertOpen() {\n      this.setState({ openAlert: true });\n    }\n  }, {\n    key: 'handleAlertClose',\n    value: function handleAlertClose() {\n      this.setState({ openAlert: false });\n    }\n  }, {\n    key: 'handleEventsOpen',\n    value: function handleEventsOpen() {\n      this.setState({ openCenterEvents: true });\n    }\n  }, {\n    key: 'handleEventsClose',\n    value: function handleEventsClose() {\n      this.setState({ openCenterEvents: false });\n    }\n  }, {\n    key: 'showEditCenterButton',\n    value: function showEditCenterButton() {\n      var isAdmin = this.props.activeUser.user.role;\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'Cancel',\n        primary: true,\n        onClick: this.handleClose\n      })];\n\n      if (isAdmin) {\n        return _react3.default.createElement(\n          'div',\n          null,\n          _react3.default.createElement(_FlatButton2.default, {\n            label: 'Edit center',\n            icon: _react3.default.createElement(_modeEdit2.default, null),\n            onClick: this.handleOpen,\n            fullWidth: true\n          }),\n          _react3.default.createElement(\n            _Dialog2.default,\n            {\n              title: 'Edit Center',\n              actions: actions,\n              modal: false,\n              open: this.state.open,\n              onRequestClose: this.handleClose,\n              autoScrollBodyContent: true,\n              style: { marginTop: '0px' }\n            },\n            _react3.default.createElement(_editCenterForm2.default, { history: this.props.history })\n          )\n        );\n      }\n    }\n  }, {\n    key: 'showCenterEventsButton',\n    value: function showCenterEventsButton(events) {\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'close',\n        primary: true,\n        onClick: this.handleEventsClose\n      })];\n\n      return _react3.default.createElement(\n        'div',\n        null,\n        _react3.default.createElement(_RaisedButton2.default, {\n          label: 'view Events',\n          onClick: this.handleEventsOpen,\n          fullWidth: true\n        }),\n        _react3.default.createElement(\n          _Dialog2.default,\n          {\n            title: 'Events Held by this center',\n            actions: actions,\n            modal: false,\n            open: this.state.openCenterEvents,\n            onRequestClose: this.handleEventsClose,\n            autoScrollBodyContent: true,\n            style: { marginTop: '-100px' }\n          },\n          _react3.default.createElement(_currentEventForCenter2.default, { event: events })\n        )\n      );\n    }\n  }, {\n    key: 'showBookCenterButton',\n    value: function showBookCenterButton() {\n      var isSignedIn = this.props.activeUser.isAuthenticated;\n      if (isSignedIn) {\n        return _react3.default.createElement(_EventModal2.default, null);\n      }\n    }\n  }, {\n    key: 'showRecommendedCenters',\n    value: function showRecommendedCenters(relatedCenterBasedOn) {\n      var isAdmin = this.props.activeUser.user.role;\n      if (!isAdmin) {\n        return _react3.default.createElement(_RecommCenter2.default, {\n          relatedCenterBasedOn: relatedCenterBasedOn,\n          fetchCenterRelatedTo: _fetchCenterRelatedTo.fetchCenterRelatedTo\n        });\n      }\n    }\n  }, {\n    key: 'deleteCenter',\n    value: function deleteCenter(id) {\n      var _this2 = this;\n\n      this.props.deleteCenterRequest(id).then(function () {\n        if (typeof _this2.props.activeCenterDetail.center === 'undefined') {\n          Materialize.toast('Center has been Deleted', 5000, 'teal');\n          _this2.props.history.push('/centers');\n        }\n      });\n    }\n  }, {\n    key: 'showAlertModal',\n    value: function showAlertModal(id) {\n      var _this3 = this;\n\n      var isAdmin = this.props.activeUser.user.role;\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'Yes',\n        primary: true,\n        onClick: function onClick() {\n          return _this3.deleteCenter(id);\n        }\n      }), _react3.default.createElement(_FlatButton2.default, {\n        label: 'No',\n        primary: true,\n        onClick: function onClick() {\n          return _this3.handleAlertClose();\n        }\n      })];\n\n      if (isAdmin) {\n        return _react3.default.createElement(\n          'div',\n          null,\n          _react3.default.createElement(_FlatButton2.default, {\n            label: 'Delete this center',\n            secondary: true,\n            icon: _react3.default.createElement(_delete2.default, null),\n            onClick: this.handleAlertOpen\n          }),\n          _react3.default.createElement(\n            _Dialog2.default,\n            {\n              actions: actions,\n              modal: false,\n              open: this.state.openAlert,\n              onRequestClose: this.handleAlertClose\n            },\n            'Are you sure you want to delete this event?'\n          )\n        );\n      }\n    }\n  }, {\n    key: 'renderFacilities',\n    value: function renderFacilities(facilities) {\n      return facilities.map(function (facility) {\n        return _react3.default.createElement(\n          'li',\n          { key: _shortid2.default.generate() },\n          facility\n        );\n      });\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _state = this.state,\n          isLoading = _state.isLoading,\n          activeCenter = _state.activeCenter;\n\n      if (activeCenter.center) {\n        var _activeCenter$center = activeCenter.center,\n            id = _activeCenter$center.id,\n            title = _activeCenter$center.title,\n            img_url = _activeCenter$center.img_url,\n            location = _activeCenter$center.location,\n            description = _activeCenter$center.description,\n            facilities = _activeCenter$center.facilities,\n            capacity = _activeCenter$center.capacity,\n            price = _activeCenter$center.price,\n            events = _activeCenter$center.events;\n\n\n        var relatedCenterBasedOn = {\n          id: id,\n          location: location,\n          facilities: facilities,\n          capacity: capacity,\n          price: price\n        };\n\n        if (events) {\n          events.map(function (event) {\n            event.center = relatedCenterBasedOn;\n          });\n        }\n\n        return _react3.default.createElement(\n          _reactDocumentTitle2.default,\n          { title: title + ' | Boots Events Manager' },\n          _react3.default.createElement(\n            'div',\n            { className: 'container' },\n            _react3.default.createElement(\n              'div',\n              { className: 'center__holdr', style: { minHeight: '100vh' } },\n              _react3.default.createElement(\n                'div',\n                { className: 'row' },\n                _react3.default.createElement(\n                  'div',\n                  { className: 'col s12 l12' },\n                  isLoading && _react3.default.createElement(_loader.CircularLoader, null),\n                  !isLoading && _react3.default.createElement(\n                    'div',\n                    { className: 'center__details', 'data-center-id': id },\n                    _react3.default.createElement(\n                      'h5',\n                      { style: { fontWeight: '500' } },\n                      title\n                    ),\n                    _react3.default.createElement(\n                      'div',\n                      { className: 'slider__holdr' },\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'carousel carousel-slider' },\n                        _react3.default.createElement(\n                          'a',\n                          { className: 'carousel-item', href: '#one' },\n                          _react3.default.createElement('img', { src: img_url, alt: title })\n                        )\n                      )\n                    ),\n                    _react3.default.createElement(\n                      'p',\n                      null,\n                      _react3.default.createElement(\n                        'i',\n                        { className: 'material-icons f15' },\n                        'location_on'\n                      ),\n                      ' ',\n                      location\n                    ),\n                    _react3.default.createElement('div', { className: 'divider' }),\n                    _react3.default.createElement(\n                      'section',\n                      null,\n                      _react3.default.createElement(\n                        'h5',\n                        null,\n                        'About this Center'\n                      ),\n                      _react3.default.createElement(\n                        'p',\n                        null,\n                        description\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Capacity'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            capacity,\n                            ' Guests'\n                          )\n                        )\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Price'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            _react3.default.createElement(\n                              'span',\n                              null,\n                              '\\u20A6',\n                              price\n                            ),\n                            ' per event'\n                          )\n                        )\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Facilities'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'ul',\n                            { className: 'facility__list' },\n                            this.renderFacilities(facilities)\n                          )\n                        )\n                      ),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s2' },\n                          this.showEditCenterButton()\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s2' },\n                          this.showAlertModal(id)\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          this.showBookCenterButton()\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s3' },\n                          this.showCenterEventsButton(events)\n                        )\n                      )\n                    )\n                  )\n                )\n              ),\n              this.showRecommendedCenters(relatedCenterBasedOn)\n            )\n          )\n        );\n      }\n      return '';\n    }\n  }]);\n\n  return CenterDetail;\n}(_react2.Component));\n\nCenterDetail.propTypes = {\n  params: _propTypes.PropTypes.object.isRequired,\n  history: _propTypes.PropTypes.object.isRequired,\n  fetchCenterRelatedTo: _propTypes.PropTypes.func.isRequired\n};\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    activeCenterDetail: state.activeCenter,\n    activeUser: state.authReducer\n  };\n};\n\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n  return (0, _redux.bindActionCreators)({\n    fetchCenterAction: _activeCenterAction.fetchCenterAction,\n    editCenterRequestAction: _activeCenterAction.editCenterRequestAction,\n    deleteCenterRequest: _deleteCenterAction.deleteCenterRequest,\n    fetchCenterRelatedTo: _fetchCenterRelatedTo.fetchCenterRelatedTo\n  }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(CenterDetail);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/centerComponent/centerDetail/centerDetail.jsx\n// module id = 1657\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/centerComponent/centerDetail/centerDetail.jsx?");
-
-/***/ })
-
-})
+webpackHotUpdate(0,{1660:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _reactReduxLoadingBar = __webpack_require__(34);\n\nvar _reactReduxLoadingBar2 = _interopRequireDefault(_reactReduxLoadingBar);\n\nvar _reactDocumentTitle = __webpack_require__(74);\n\nvar _reactDocumentTitle2 = _interopRequireDefault(_reactDocumentTitle);\n\nvar _fixedNav = __webpack_require__(188);\n\nvar _fixedNav2 = _interopRequireDefault(_fixedNav);\n\nvar _centerDetail = __webpack_require__(1661);\n\nvar _centerDetail2 = _interopRequireDefault(_centerDetail);\n\nvar _FloatingActionButton = __webpack_require__(127);\n\nvar _FloatingActionButton2 = _interopRequireDefault(_FloatingActionButton);\n\nvar _footer = __webpack_require__(184);\n\nvar _footer2 = _interopRequireDefault(_footer);\n\nvar _modals = __webpack_require__(128);\n\nvar _modals2 = _interopRequireDefault(_modals);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  CenterDetailIndex: {\n    displayName: 'CenterDetailIndex'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/index.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/index.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar CenterDetailIndex = _wrapComponent('CenterDetailIndex')(function (_Component) {\n  _inherits(CenterDetailIndex, _Component);\n\n  function CenterDetailIndex() {\n    _classCallCheck(this, CenterDetailIndex);\n\n    return _possibleConstructorReturn(this, (CenterDetailIndex.__proto__ || Object.getPrototypeOf(CenterDetailIndex)).apply(this, arguments));\n  }\n\n  _createClass(CenterDetailIndex, [{\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      $(\".button-collapse\").sideNav();\n    }\n  }, {\n    key: 'showSandWichNavBar',\n    value: function showSandWichNavBar() {\n      return _react3.default.createElement(\n        'div',\n        null,\n        _react3.default.createElement(\n          'ul',\n          { id: 'slide-out', className: 'sidenav' },\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              'div',\n              { className: 'user-view' },\n              _react3.default.createElement(\n                'div',\n                { className: 'background' },\n                _react3.default.createElement('img', { src: 'images/office.jpg', alt: '' })\n              ),\n              _react3.default.createElement(\n                'a',\n                { href: '#user' },\n                _react3.default.createElement('img', { className: 'circle', src: 'images/yuna.jpg' })\n              ),\n              _react3.default.createElement(\n                'a',\n                { href: '#name' },\n                _react3.default.createElement(\n                  'span',\n                  { className: 'white-text name' },\n                  'John Doe'\n                )\n              ),\n              _react3.default.createElement(\n                'a',\n                { href: '#email' },\n                _react3.default.createElement(\n                  'span',\n                  { className: 'white-text email' },\n                  'jdandturk@gmail.com'\n                )\n              )\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              'a',\n              { href: '#!' },\n              _react3.default.createElement(\n                'i',\n                { className: 'material-icons' },\n                'cloud'\n              ),\n              'First Link With Icon'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              'a',\n              { href: '#!' },\n              'Second Link'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement('div', { className: 'divider' })\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              'a',\n              { className: 'subheader' },\n              'Subheader'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              'a',\n              { className: 'waves-effect', href: '#!' },\n              'Third Link With Waves'\n            )\n          )\n        )\n      );\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      return _react3.default.createElement(\n        _reactDocumentTitle2.default,\n        { title: 'Center Details | Boots Events Manager' },\n        _react3.default.createElement(\n          'div',\n          null,\n          _react3.default.createElement(_reactReduxLoadingBar2.default, { style: { backgroundImage: 'linear-gradient(to top left, rgba(72, 132, 179, 0.7), rgba(144, 236, 146, 0.7))', height: 2 + 'px' } }),\n          _react3.default.createElement(\n            'div',\n            { className: 'body__holdr' },\n            this.showSandWichNavBar(),\n            _react3.default.createElement(_fixedNav2.default, null),\n            _react3.default.createElement(_centerDetail2.default, { params: this.props.match.params, history: this.props.history }),\n            _react3.default.createElement(_FloatingActionButton2.default, null),\n            _react3.default.createElement(_footer2.default, null)\n          ),\n          _react3.default.createElement(_modals2.default, null)\n        )\n      );\n    }\n  }]);\n\n  return CenterDetailIndex;\n}(_react2.Component));\n\nexports.default = CenterDetailIndex;\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/centerComponent/centerDetail/index.jsx\n// module id = 1660\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/centerComponent/centerDetail/index.jsx?")},1661:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _propTypes = __webpack_require__(4);\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _redux = __webpack_require__(23);\n\nvar _shortid = __webpack_require__(35);\n\nvar _shortid2 = _interopRequireDefault(_shortid);\n\nvar _reactDocumentTitle = __webpack_require__(74);\n\nvar _reactDocumentTitle2 = _interopRequireDefault(_reactDocumentTitle);\n\nvar _Dialog = __webpack_require__(91);\n\nvar _Dialog2 = _interopRequireDefault(_Dialog);\n\nvar _RaisedButton = __webpack_require__(182);\n\nvar _RaisedButton2 = _interopRequireDefault(_RaisedButton);\n\nvar _FlatButton = __webpack_require__(47);\n\nvar _FlatButton2 = _interopRequireDefault(_FlatButton);\n\nvar _modeEdit = __webpack_require__(289);\n\nvar _modeEdit2 = _interopRequireDefault(_modeEdit);\n\nvar _delete = __webpack_require__(288);\n\nvar _delete2 = _interopRequireDefault(_delete);\n\nvar _loader = __webpack_require__(90);\n\nvar _activeCenterAction = __webpack_require__(1662);\n\nvar _deleteCenterAction = __webpack_require__(1663);\n\nvar _currentEventForCenter = __webpack_require__(1664);\n\nvar _currentEventForCenter2 = _interopRequireDefault(_currentEventForCenter);\n\nvar _RecommCenter = __webpack_require__(1665);\n\nvar _RecommCenter2 = _interopRequireDefault(_RecommCenter);\n\nvar _EventModal = __webpack_require__(1666);\n\nvar _EventModal2 = _interopRequireDefault(_EventModal);\n\nvar _editCenterForm = __webpack_require__(1667);\n\nvar _editCenterForm2 = _interopRequireDefault(_editCenterForm);\n\nvar _fetchCenterRelatedTo = __webpack_require__(1668);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  CenterDetail: {\n    displayName: 'CenterDetail'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/centerDetail.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/centerDetail.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar CenterDetail = _wrapComponent('CenterDetail')(function (_Component) {\n  _inherits(CenterDetail, _Component);\n\n  function CenterDetail(props) {\n    _classCallCheck(this, CenterDetail);\n\n    var _this = _possibleConstructorReturn(this, (CenterDetail.__proto__ || Object.getPrototypeOf(CenterDetail)).call(this, props));\n\n    _this.state = {\n      isLoading: true,\n      openAlert: false,\n      open: false,\n      openCenterEvents: false,\n      activeCenter: {\n        centr: {\n          title: 'center'\n        }\n      }\n    };\n\n    _this.handleOpen = _this.handleOpen.bind(_this);\n    _this.handleClose = _this.handleClose.bind(_this);\n    _this.handleAlertOpen = _this.handleAlertOpen.bind(_this);\n    _this.handleAlertClose = _this.handleAlertClose.bind(_this);\n    _this.handleEventsOpen = _this.handleEventsOpen.bind(_this);\n    _this.handleEventsClose = _this.handleEventsClose.bind(_this);\n    return _this;\n  }\n\n  _createClass(CenterDetail, [{\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      $('.modal').modal('close');\n      $('.tooltipped').tooltip({ delay: 50 });\n      var params = this.props.params;\n\n      this.props.fetchCenterAction(params.id);\n    }\n  }, {\n    key: 'componentWillReceiveProps',\n    value: function componentWillReceiveProps(newProps) {\n      if (this.props.params.id !== newProps.params.id) {\n        newProps.fetchCenterAction(newProps.params.id);\n      }\n\n      if (typeof newProps.activeCenterDetail.center !== 'undefined') {\n        newProps.activeCenterDetail.center.events = newProps.activeCenterDetail.events;\n        if (newProps.activeCenterDetail.center.events) {\n          delete newProps.activeCenterDetail.events;\n        }\n        this.setState({ isLoading: false, activeCenter: newProps.activeCenterDetail });\n      }\n    }\n  }, {\n    key: 'editCenter',\n    value: function editCenter() {\n      this.props.editCenterRequestAction();\n    }\n  }, {\n    key: 'handleOpen',\n    value: function handleOpen() {\n      this.props.editCenterRequestAction();\n      this.setState({ open: true });\n    }\n  }, {\n    key: 'handleClose',\n    value: function handleClose() {\n      this.setState({ open: false });\n    }\n  }, {\n    key: 'handleAlertOpen',\n    value: function handleAlertOpen() {\n      this.setState({ openAlert: true });\n    }\n  }, {\n    key: 'handleAlertClose',\n    value: function handleAlertClose() {\n      this.setState({ openAlert: false });\n    }\n  }, {\n    key: 'handleEventsOpen',\n    value: function handleEventsOpen() {\n      this.setState({ openCenterEvents: true });\n    }\n  }, {\n    key: 'handleEventsClose',\n    value: function handleEventsClose() {\n      this.setState({ openCenterEvents: false });\n    }\n  }, {\n    key: 'showEditCenterButton',\n    value: function showEditCenterButton() {\n      var isAdmin = this.props.activeUser.user.role;\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'Cancel',\n        primary: true,\n        onClick: this.handleClose\n      })];\n\n      if (isAdmin) {\n        return _react3.default.createElement(\n          'div',\n          null,\n          _react3.default.createElement(_FlatButton2.default, {\n            label: 'Edit center',\n            icon: _react3.default.createElement(_modeEdit2.default, null),\n            onClick: this.handleOpen,\n            fullWidth: true\n          }),\n          _react3.default.createElement(\n            _Dialog2.default,\n            {\n              title: 'Edit Center',\n              actions: actions,\n              modal: false,\n              open: this.state.open,\n              onRequestClose: this.handleClose,\n              autoScrollBodyContent: true,\n              style: { marginTop: '0px' }\n            },\n            _react3.default.createElement(_editCenterForm2.default, { history: this.props.history })\n          )\n        );\n      }\n    }\n  }, {\n    key: 'showCenterEventsButton',\n    value: function showCenterEventsButton(events) {\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'close',\n        primary: true,\n        onClick: this.handleEventsClose\n      })];\n\n      return _react3.default.createElement(\n        'div',\n        null,\n        _react3.default.createElement(_RaisedButton2.default, {\n          label: 'view Events',\n          onClick: this.handleEventsOpen,\n          fullWidth: true\n        }),\n        _react3.default.createElement(\n          _Dialog2.default,\n          {\n            title: 'Events Held by this center',\n            actions: actions,\n            modal: false,\n            open: this.state.openCenterEvents,\n            onRequestClose: this.handleEventsClose,\n            autoScrollBodyContent: true,\n            style: { marginTop: '-100px' }\n          },\n          _react3.default.createElement(_currentEventForCenter2.default, { event: events })\n        )\n      );\n    }\n  }, {\n    key: 'showBookCenterButton',\n    value: function showBookCenterButton() {\n      var isSignedIn = this.props.activeUser.isAuthenticated;\n      if (isSignedIn) {\n        return _react3.default.createElement(_EventModal2.default, null);\n      }\n    }\n  }, {\n    key: 'showRecommendedCenters',\n    value: function showRecommendedCenters(relatedCenterBasedOn) {\n      var isAdmin = this.props.activeUser.user.role;\n      if (!isAdmin) {\n        return _react3.default.createElement(_RecommCenter2.default, {\n          relatedCenterBasedOn: relatedCenterBasedOn,\n          fetchCenterRelatedTo: _fetchCenterRelatedTo.fetchCenterRelatedTo\n        });\n      }\n    }\n  }, {\n    key: 'deleteCenter',\n    value: function deleteCenter(id) {\n      var _this2 = this;\n\n      this.props.deleteCenterRequest(id).then(function () {\n        if (typeof _this2.props.activeCenterDetail.center === 'undefined') {\n          Materialize.toast('Center has been Deleted', 5000, 'teal');\n          _this2.props.history.push('/centers');\n        }\n      });\n    }\n  }, {\n    key: 'showAlertModal',\n    value: function showAlertModal(id) {\n      var _this3 = this;\n\n      var isAdmin = this.props.activeUser.user.role;\n      var actions = [_react3.default.createElement(_FlatButton2.default, {\n        label: 'Yes',\n        primary: true,\n        onClick: function onClick() {\n          return _this3.deleteCenter(id);\n        }\n      }), _react3.default.createElement(_FlatButton2.default, {\n        label: 'No',\n        primary: true,\n        onClick: function onClick() {\n          return _this3.handleAlertClose();\n        }\n      })];\n\n      if (isAdmin) {\n        return _react3.default.createElement(\n          'div',\n          null,\n          _react3.default.createElement(_FlatButton2.default, {\n            label: 'Delete this center',\n            secondary: true,\n            icon: _react3.default.createElement(_delete2.default, null),\n            onClick: this.handleAlertOpen\n          }),\n          _react3.default.createElement(\n            _Dialog2.default,\n            {\n              actions: actions,\n              modal: false,\n              open: this.state.openAlert,\n              onRequestClose: this.handleAlertClose\n            },\n            'Are you sure you want to delete this event?'\n          )\n        );\n      }\n    }\n  }, {\n    key: 'renderFacilities',\n    value: function renderFacilities(facilities) {\n      return facilities.map(function (facility) {\n        return _react3.default.createElement(\n          'li',\n          { key: _shortid2.default.generate() },\n          facility\n        );\n      });\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _state = this.state,\n          isLoading = _state.isLoading,\n          activeCenter = _state.activeCenter;\n\n      if (activeCenter.center) {\n        var _activeCenter$center = activeCenter.center,\n            id = _activeCenter$center.id,\n            title = _activeCenter$center.title,\n            img_url = _activeCenter$center.img_url,\n            location = _activeCenter$center.location,\n            description = _activeCenter$center.description,\n            facilities = _activeCenter$center.facilities,\n            capacity = _activeCenter$center.capacity,\n            price = _activeCenter$center.price,\n            events = _activeCenter$center.events;\n\n\n        var relatedCenterBasedOn = {\n          id: id,\n          location: location,\n          facilities: facilities,\n          capacity: capacity,\n          price: price\n        };\n\n        if (events) {\n          events.map(function (event) {\n            event.center = relatedCenterBasedOn;\n          });\n        }\n\n        return _react3.default.createElement(\n          _reactDocumentTitle2.default,\n          { title: title + ' | Boots Events Manager' },\n          _react3.default.createElement(\n            'div',\n            { className: 'container' },\n            _react3.default.createElement(\n              'div',\n              { className: 'center__holdr', style: { minHeight: '100vh' } },\n              _react3.default.createElement(\n                'div',\n                { className: 'row' },\n                _react3.default.createElement(\n                  'div',\n                  { className: 'col s12 l12' },\n                  isLoading && _react3.default.createElement(_loader.CircularLoader, null),\n                  !isLoading && _react3.default.createElement(\n                    'div',\n                    { className: 'center__details', 'data-center-id': id },\n                    _react3.default.createElement(\n                      'h5',\n                      { style: { fontWeight: '500' } },\n                      title\n                    ),\n                    _react3.default.createElement(\n                      'div',\n                      { className: 'slider__holdr' },\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'carousel carousel-slider' },\n                        _react3.default.createElement(\n                          'a',\n                          { className: 'carousel-item', href: '#one' },\n                          _react3.default.createElement('img', { src: img_url, alt: title })\n                        )\n                      )\n                    ),\n                    _react3.default.createElement(\n                      'p',\n                      null,\n                      _react3.default.createElement(\n                        'i',\n                        { className: 'material-icons f15' },\n                        'location_on'\n                      ),\n                      ' ',\n                      location\n                    ),\n                    _react3.default.createElement('div', { className: 'divider' }),\n                    _react3.default.createElement(\n                      'section',\n                      null,\n                      _react3.default.createElement(\n                        'h5',\n                        null,\n                        'About this Center'\n                      ),\n                      _react3.default.createElement(\n                        'p',\n                        null,\n                        description\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Capacity'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            capacity,\n                            ' Guests'\n                          )\n                        )\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Price'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            _react3.default.createElement(\n                              'span',\n                              null,\n                              '\\u20A6',\n                              price\n                            ),\n                            ' per event'\n                          )\n                        )\n                      ),\n                      _react3.default.createElement('div', { className: 'divider' }),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          _react3.default.createElement(\n                            'p',\n                            null,\n                            'Facilities'\n                          )\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s8' },\n                          _react3.default.createElement(\n                            'ul',\n                            { className: 'facility__list' },\n                            this.renderFacilities(facilities)\n                          )\n                        )\n                      ),\n                      _react3.default.createElement(\n                        'div',\n                        { className: 'row' },\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s2' },\n                          this.showEditCenterButton()\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s2' },\n                          this.showAlertModal(id)\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s4' },\n                          this.showBookCenterButton()\n                        ),\n                        _react3.default.createElement(\n                          'div',\n                          { className: 'col s3' },\n                          this.showCenterEventsButton(events)\n                        )\n                      )\n                    )\n                  )\n                )\n              ),\n              this.showRecommendedCenters(relatedCenterBasedOn)\n            )\n          )\n        );\n      }\n      return '';\n    }\n  }]);\n\n  return CenterDetail;\n}(_react2.Component));\n\nCenterDetail.propTypes = {\n  params: _propTypes.PropTypes.object.isRequired,\n  history: _propTypes.PropTypes.object.isRequired,\n  fetchCenterRelatedTo: _propTypes.PropTypes.func.isRequired\n};\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    activeCenterDetail: state.activeCenter,\n    activeUser: state.authReducer\n  };\n};\n\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n  return (0, _redux.bindActionCreators)({\n    fetchCenterAction: _activeCenterAction.fetchCenterAction,\n    editCenterRequestAction: _activeCenterAction.editCenterRequestAction,\n    deleteCenterRequest: _deleteCenterAction.deleteCenterRequest,\n    fetchCenterRelatedTo: _fetchCenterRelatedTo.fetchCenterRelatedTo\n  }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(CenterDetail);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/centerComponent/centerDetail/centerDetail.jsx\n// module id = 1661\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/centerComponent/centerDetail/centerDetail.jsx?")},1665:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _propTypes = __webpack_require__(4);\n\nvar _reactRouterDom = __webpack_require__(27);\n\nvar _shortid = __webpack_require__(35);\n\nvar _shortid2 = _interopRequireDefault(_shortid);\n\nvar _isEmpty = __webpack_require__(42);\n\nvar _isEmpty2 = _interopRequireDefault(_isEmpty);\n\nvar _loader = __webpack_require__(90);\n\nvar _helpers = __webpack_require__(129);\n\nvar _helpers2 = _interopRequireDefault(_helpers);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  RecommCenter: {\n    displayName: 'RecommCenter'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/RecommCenter.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/centerComponent/centerDetail/RecommCenter.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar RecommCenter = _wrapComponent('RecommCenter')(function (_Component) {\n  _inherits(RecommCenter, _Component);\n\n  function RecommCenter(props) {\n    _classCallCheck(this, RecommCenter);\n\n    var _this = _possibleConstructorReturn(this, (RecommCenter.__proto__ || Object.getPrototypeOf(RecommCenter)).call(this, props));\n\n    _this.helper = new _helpers2.default();\n    _this.state = {\n      isLoading: true,\n      error: false,\n      noCenter: 'There are no related centers',\n      errorMessage: '',\n      relatedCenters: [],\n      currentCenterId: _this.props.relatedCenterBasedOn.id\n    };\n    return _this;\n  }\n\n  _createClass(RecommCenter, [{\n    key: 'fetchCenter',\n    value: function fetchCenter(relatedCenters) {\n      var _this2 = this;\n\n      this.props.fetchCenterRelatedTo(relatedCenters).then(function (_ref) {\n        var data = _ref.data;\n\n        _this2.setState({ isLoading: false, relatedCenters: data.centers });\n      }).catch(function () {\n        _this2.setState({ isLoading: false, error: true, errorMessage: _this2.state.noCenter });\n      });\n    }\n  }, {\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      this.fetchCenter(this.props.relatedCenterBasedOn);\n    }\n  }, {\n    key: 'componentWillReceiveProps',\n    value: function componentWillReceiveProps(newProps) {\n      this.fetchCenter(newProps.relatedCenterBasedOn);\n    }\n  }, {\n    key: 'sortAndShowRecommended',\n    value: function sortAndShowRecommended() {\n      var _this3 = this;\n\n      if (!(0, _isEmpty2.default)(this.state.relatedCenters)) {\n        return this.state.relatedCenters.map(function (center, index) {\n          var to = '/center/' + center.id + '/' + _this3.helper.sanitizeString(center.title);\n          return _react3.default.createElement(\n            'div',\n            { key: _shortid2.default.generate(), className: 'col s12 m6 l4' },\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: to, href: to },\n              _react3.default.createElement(\n                'div',\n                { className: 'card' },\n                !!center.img_url && _react3.default.createElement(\n                  'div',\n                  { className: 'card-image center__image' },\n                  _react3.default.createElement('img', { src: center.img_url, alt: center.title })\n                ),\n                _react3.default.createElement(\n                  'div',\n                  { className: 'card-content black-text' },\n                  _react3.default.createElement(\n                    'p',\n                    { className: 'bold' },\n                    center.title\n                  ),\n                  _react3.default.createElement(\n                    'p',\n                    { className: 'light__font' },\n                    _react3.default.createElement(\n                      'i',\n                      { className: 'material-icons f15' },\n                      'location_on'\n                    ),\n                    center.location\n                  )\n                )\n              )\n            )\n          );\n        });\n      }\n      return _react3.default.createElement(\n        'p',\n        null,\n        this.state.noCenter\n      );\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _state = this.state,\n          isLoading = _state.isLoading,\n          error = _state.error,\n          errorMessage = _state.errorMessage;\n\n      var eachCenter = this.sortAndShowRecommended();\n      return _react3.default.createElement(\n        'div',\n        { className: 'row' },\n        _react3.default.createElement('div', { className: 'divider' }),\n        _react3.default.createElement(\n          'h5',\n          null,\n          'Recommended Center'\n        ),\n        isLoading ? _react3.default.createElement(_loader.CircularLoader, null) : _react3.default.createElement(\n          'div',\n          { className: 'row' },\n          error ? errorMessage : (0, _isEmpty2.default)(eachCenter) ? this.state.noCenter : eachCenter\n        )\n      );\n    }\n  }]);\n\n  return RecommCenter;\n}(_react2.Component));\n\nRecommCenter.propTypes = {\n  relatedCenterBasedOn: _propTypes.PropTypes.object.isRequired,\n  fetchCenterRelatedTo: _propTypes.PropTypes.func.isRequired\n};\n\nexports.default = RecommCenter;\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/centerComponent/centerDetail/RecommCenter.jsx\n// module id = 1665\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/centerComponent/centerDetail/RecommCenter.jsx?")},1666:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _redux = __webpack_require__(23);\n\nvar _eventsActions = __webpack_require__(72);\n\nvar _formInput = __webpack_require__(73);\n\nvar _formInput2 = _interopRequireDefault(_formInput);\n\nvar _validateInput = __webpack_require__(126);\n\nvar _actions = __webpack_require__(28);\n\nvar _Dialog = __webpack_require__(91);\n\nvar _Dialog2 = _interopRequireDefault(_Dialog);\n\nvar _FlatButton = __webpack_require__(47);\n\nvar _FlatButton2 = _interopRequireDefault(_FlatButton);\n\nvar _RaisedButton = __webpack_require__(182);\n\nvar _RaisedButton2 = _interopRequireDefault(_RaisedButton);\n\nvar _DatePicker = __webpack_require__(294);\n\nvar _DatePicker2 = _interopRequireDefault(_DatePicker);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n    EventModal: {\n        displayName: 'EventModal'\n    }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n    filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/modals/EventModal.jsx',\n    components: _components,\n    locals: [module],\n    imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n    filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/modals/EventModal.jsx',\n    components: _components,\n    locals: [],\n    imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n    return function (Component) {\n        return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n    };\n}\n\nvar EventModal = _wrapComponent('EventModal')(function (_Component) {\n    _inherits(EventModal, _Component);\n\n    function EventModal(props) {\n        _classCallCheck(this, EventModal);\n\n        var _this = _possibleConstructorReturn(this, (EventModal.__proto__ || Object.getPrototypeOf(EventModal)).call(this, props));\n\n        _this.handleChangeStartDate = function (e, date) {\n            if (new Date(date) < new Date()) {\n                Materialize.toast('Date isn\\'t correct. Should be a day after today not before', 5000, 'red');\n                _this.setState({\n                    startDate: {}\n                });\n            } else {\n                _this.setState({\n                    startDate: date.toDateString()\n                });\n            }\n        };\n\n        _this.handleChangeEndDate = function (e, date) {\n            if (new Date(date) < new Date()) {\n                Materialize.toast('Date isn\\'t correct. Should be a day after today not before', 5000, 'red');\n                _this.setState({\n                    endDate: {}\n                });\n            } else {\n                _this.setState({\n                    endDate: date.toDateString()\n                });\n            }\n        };\n\n        _this.handleOpen = function () {\n            _this.setState({ open: true });\n        };\n\n        _this.handleClose = function () {\n            _this.setState({ open: false });\n        };\n\n        var startDate = new Date();\n        var endDate = new Date();\n        startDate.setFullYear(startDate.getFullYear() - 1);\n        endDate.setFullYear(endDate.getFullYear() - 1);\n\n        _this.state = {\n            open: false,\n            disableYearSelection: false,\n            isLoading: false,\n            errors: {},\n            editEvent: false,\n            centerId: 0,\n            userId: 0,\n            title: '',\n            img_url: '',\n            startDate: null,\n            endDate: null,\n            description: ''\n        };\n\n        _this.handleInputChange = _this.handleInputChange.bind(_this);\n        _this.handleEventSubmit = _this.handleEventSubmit.bind(_this);\n        _this.onFileChange = _this.onFileChange.bind(_this);\n        return _this;\n    }\n\n    _createClass(EventModal, [{\n        key: 'updateProps',\n        value: function updateProps(newProps) {\n            if (newProps.editEvent) {\n                var _newProps$eventToEdit = newProps.eventToEdit,\n                    title = _newProps$eventToEdit.title,\n                    img_url = _newProps$eventToEdit.img_url,\n                    startDate = _newProps$eventToEdit.startDate,\n                    endDate = _newProps$eventToEdit.endDate,\n                    description = _newProps$eventToEdit.description;\n\n                this.setState({\n                    editEvent: true,\n                    centerId: newProps.activeCenter.center.id,\n                    userId: newProps.actUser.user.id,\n                    title: title,\n                    img_url: img_url,\n                    startDate: startDate,\n                    endDate: endDate,\n                    description: description\n                });\n            } else {\n                this.setState({\n                    centerId: newProps.activeCenter.center.id,\n                    userId: newProps.actUser.user.id\n                });\n            }\n        }\n    }, {\n        key: 'componentDidMount',\n        value: function componentDidMount() {\n            this.updateProps(this.props);\n        }\n    }, {\n        key: 'isValid',\n        value: function isValid() {\n            var _validateEventInput = (0, _validateInput.validateEventInput)(this.state),\n                errors = _validateEventInput.errors,\n                isValid = _validateEventInput.isValid;\n\n            if (!isValid) {\n                this.setState({ errors: errors });\n            }\n            return isValid;\n        }\n    }, {\n        key: 'handleInputChange',\n        value: function handleInputChange(e) {\n            if (!!this.state.errors[e.target.name]) {\n                var _setState;\n\n                var errors = Object.assign({}, !!this.state.errors);\n                delete errors[e.target.name];\n                this.setState((_setState = {}, _defineProperty(_setState, e.target.name, e.target.value), _defineProperty(_setState, 'errors', errors), _setState));\n            } else {\n                this.setState(_defineProperty({}, e.target.name, e.target.value));\n            }\n        }\n    }, {\n        key: 'onFileChange',\n        value: function onFileChange(e) {\n            var file = e.target.files[0];\n            if (file.type.indexOf('image/') > -1) {\n                // only image file\n                if (file.size < 2000000) {\n                    this.setState({\n                        img_url: file\n                    });\n                } else {\n                    Materialize.toast('File too large', 5000, 'red');\n                }\n            } else {\n                Materialize.toast('Image files only please', 5000, 'red');\n            }\n        }\n    }, {\n        key: 'handleEventSubmit',\n        value: function handleEventSubmit(e) {\n            e.preventDefault();\n            if (this.isValid()) {\n                this.setState({\n                    isLoading: true\n                });\n                this.props.createEventRequest(this.state);\n                // .then((data)=>{\n                //     console.log('res from action', data);\n                //     this.setState({isLoading: false});\n                //     if(data.type === ADD_EVENT){\n                //         Materialize.toast('Event has been created successfully', 5000, 'teal');\n                //         this.setState({ title: '', description: '' });\n                //         this.handleClose();\n                //     }else{\n                //         Materialize.toast(data.message, 5000, 'red');\n                //     }\n                // });\n            }\n        }\n    }, {\n        key: 'render',\n        value: function render() {\n            var _state = this.state,\n                isLoading = _state.isLoading,\n                editEvent = _state.editEvent,\n                title = _state.title,\n                description = _state.description,\n                endDate = _state.endDate,\n                startDate = _state.startDate,\n                errors = _state.errors;\n\n            var actions = [_react3.default.createElement(_FlatButton2.default, {\n                label: 'Cancel',\n                primary: true,\n                onClick: this.handleClose\n            }), _react3.default.createElement(_FlatButton2.default, {\n                label: isLoading ? _react3.default.createElement('img', { style: { marginTop: \"10px\" }, src: '/image/loader/loading.gif' }) : 'Add Event',\n                primary: true,\n                keyboardFocused: true,\n                onClick: this.handleEventSubmit\n            })];\n            return _react3.default.createElement(\n                'div',\n                null,\n                _react3.default.createElement(_RaisedButton2.default, {\n                    label: 'Book this center',\n                    primary: true,\n                    onClick: this.handleOpen\n                }),\n                _react3.default.createElement(\n                    _Dialog2.default,\n                    {\n                        title: editEvent ? 'Edit Event' : 'Create Event',\n                        actions: actions,\n                        modal: false,\n                        open: this.state.open,\n                        onRequestClose: this.handleClose,\n                        autoScrollBodyContent: true,\n                        style: { marginTop: '0px' }\n                    },\n                    _react3.default.createElement(\n                        'div',\n                        { className: 'row', style: { marginTop: '20px' } },\n                        _react3.default.createElement(\n                            'form',\n                            { className: 'col s12', id: 'add-event-form' },\n                            _react3.default.createElement(\n                                'div',\n                                { className: 'row' },\n                                _react3.default.createElement(\n                                    'div',\n                                    { className: 'col s6' },\n                                    _react3.default.createElement(\n                                        'div',\n                                        { className: 'file-field input-field' },\n                                        _react3.default.createElement(\n                                            'div',\n                                            { className: 'btn' },\n                                            _react3.default.createElement(\n                                                'span',\n                                                null,\n                                                'Upload'\n                                            ),\n                                            _react3.default.createElement('input', {\n                                                type: 'file',\n                                                name: 'img_url',\n                                                accept: 'image/jpeg,jpg,png,gif',\n                                                onChange: this.onFileChange\n                                            })\n                                        ),\n                                        _react3.default.createElement(\n                                            'div',\n                                            { className: 'file-path-wrapper' },\n                                            _react3.default.createElement('input', { className: 'file-path validate', type: 'text',\n                                                placeholder: 'Upload an image here' })\n                                        )\n                                    )\n                                ),\n                                _react3.default.createElement(\n                                    'div',\n                                    { className: 'input-field col s6' },\n                                    _react3.default.createElement(_formInput2.default, {\n                                        type: 'text',\n                                        fieldId: 'event_title',\n                                        nameField: 'title',\n                                        label: 'Title',\n                                        value: title,\n                                        error: errors.title || '',\n                                        onChange: this.handleInputChange\n                                    })\n                                )\n                            ),\n                            _react3.default.createElement(\n                                'div',\n                                { className: 'row' },\n                                _react3.default.createElement(\n                                    'div',\n                                    { className: 'input-field col s6' },\n                                    _react3.default.createElement(_DatePicker2.default, {\n                                        onChange: this.handleChangeStartDate,\n                                        autoOk: true,\n                                        floatingLabelText: 'Start Date',\n                                        disableYearSelection: this.state.disableYearSelection\n                                    }),\n                                    errors.startDate && _react3.default.createElement(\n                                        'span',\n                                        { className: 'red-text accent-1' },\n                                        errors.startDate\n                                    )\n                                ),\n                                _react3.default.createElement(\n                                    'div',\n                                    { className: 'input-field col s6' },\n                                    _react3.default.createElement(_DatePicker2.default, {\n                                        onChange: this.handleChangeEndDate,\n                                        autoOk: true,\n                                        floatingLabelText: 'End Date',\n                                        disableYearSelection: this.state.disableYearSelection\n                                    }),\n                                    errors.endDate && _react3.default.createElement(\n                                        'span',\n                                        { className: 'red-text accent-1' },\n                                        errors.endDate\n                                    )\n                                )\n                            ),\n                            _react3.default.createElement(\n                                'div',\n                                { className: 'row' },\n                                _react3.default.createElement(\n                                    'div',\n                                    { className: 'input-field col s12' },\n                                    _react3.default.createElement(\n                                        'label',\n                                        { htmlFor: 'description' },\n                                        'Description'\n                                    ),\n                                    _react3.default.createElement('textarea', {\n                                        id: 'description',\n                                        type: 'text',\n                                        name: 'description',\n                                        className: 'materialize-textarea validate',\n                                        required: true,\n                                        onChange: this.handleInputChange,\n                                        value: description\n                                    })\n                                )\n                            )\n                        )\n                    )\n                )\n            );\n        }\n    }]);\n\n    return EventModal;\n}(_react2.Component));\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        activeCenter: state.activeCenter,\n        event: state.eventReducer,\n        actUser: state.authReducer\n    };\n};\n\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n    return (0, _redux.bindActionCreators)({ createEventRequest: _eventsActions.createEventRequest }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(EventModal);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/modals/EventModal.jsx\n// module id = 1666\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/modals/EventModal.jsx?")},179:!1,1791:function(module,exports,__webpack_require__){"use strict";eval('/* WEBPACK VAR INJECTION */(function(process) {\n\nObject.defineProperty(exports, "__esModule", {\n  value: true\n});\n\nvar _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };\n\nvar _isEmpty = __webpack_require__(42);\n\nvar _isEmpty2 = _interopRequireDefault(_isEmpty);\n\nvar _actions = __webpack_require__(28);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nvar pageLimit = process.env.DATA_LIMIT;\nvar newState = void 0;\n\nexports.default = function () {\n  var state = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n  var action = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};\n\n  switch (action.type) {\n    case _actions.FETCH_EVENTS:\n      return action.payload;\n\n    case _actions.ADD_EVENT:\n      if (state.events) {\n        state.events.unshift(action.payload);\n        state.totalCount = state.events.length;\n        state.pageCount = Math.ceil(state.totalCount / pageLimit);\n        return state;\n      }\n      state.events = [];\n      state.events.unshift(action.payload);\n      state.totalCount = state.events.length;\n      state.pageCount = Math.ceil(state.totalCount / pageLimit);\n      return state;\n\n    case _actions.EDIT_EVENT_REQUEST:\n      return _extends({}, state, {\n        eventToEdit: action.payload,\n        editEvent: true\n      });\n\n    case _actions.EDIT_EVENT:\n      newState = Object.assign({}, state);\n      newState.events.map(function (event, index) {\n        if (event.id === action.payload.id) {\n          newState.events[index] = action.payload;\n        }\n      });\n      newState.totalCount = newState.events.length;\n      newState.pageSize = newState.totalCount;\n      newState.pageCount = Math.ceil(newState.totalCount / pageLimit);\n      return newState;\n\n    case _actions.REMOVE_EVENT:\n      newState = Object.assign({}, state);\n      newState.events.map(function (event, index) {\n        if (event.id === action.payload.id) {\n          delete newState.events[index];\n        }\n      });\n      newState.totalCount = newState.events.length;\n      newState.pageSize = newState.totalCount;\n      newState.pageCount = Math.ceil(newState.totalCount / pageLimit);\n      return newState;\n\n    case _actions.SESSION_EVENTS:\n      newState = Object.assign({}, state);\n      newState.sessEvents = action.payload;\n      return newState;\n\n    case _actions.LOADMORE_EVENT_FAILURE:\n      return _extends({}, state, { loadingmore: false });\n\n    case _actions.LOADMORE_EVENT_REQUEST:\n      return _extends({}, state, { loadmore: true, loadingmore: true });\n\n    case _actions.LOADMORE_EVENT_SUCCESS:\n      newState = Object.assign({}, state);\n      newState.events = newState.events.concat(action.payload);\n      newState.loadingmore = false;\n      newState.page = parseInt(newState.page + 1, 10);\n      newState.pageSize = parseInt(newState.pageSize + action.payload.length, 10);\n      if (newState.pageSize === newState.totalCount) {\n        newState.loadmore = false;\n      }\n      return newState;\n\n    case _actions.SEARCH_EVENT_TITLE:\n      newState = Object.assign({}, state);\n      if (!(0, _isEmpty2.default)(action.events)) {\n        newState.events = action.events;\n      } else {\n        newState.events = state.events;\n      }\n      return newState;\n\n    case _actions.SEARCH_EVENT_TITLE_FAILED:\n      newState = Object.assign({}, state);\n      newState.searchFailure = true;\n      return newState;\n\n    default:\n      return state;\n  }\n};\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(3)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/reducers/eventReducer.js\n// module id = 1791\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/reducers/eventReducer.js?')},188:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _redux = __webpack_require__(23);\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _reactRouterDom = __webpack_require__(27);\n\nvar _IconMenu = __webpack_require__(116);\n\nvar _IconMenu2 = _interopRequireDefault(_IconMenu);\n\nvar _IconButton = __webpack_require__(54);\n\nvar _IconButton2 = _interopRequireDefault(_IconButton);\n\nvar _moreVert = __webpack_require__(89);\n\nvar _moreVert2 = _interopRequireDefault(_moreVert);\n\nvar _MenuItem = __webpack_require__(55);\n\nvar _MenuItem2 = _interopRequireDefault(_MenuItem);\n\nvar _Divider = __webpack_require__(178);\n\nvar _Divider2 = _interopRequireDefault(_Divider);\n\nvar _authActions = __webpack_require__(56);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  FixedNav: {\n    displayName: 'FixedNav'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/fixedNav.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/fixedNav.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar FixedNav = _wrapComponent('FixedNav')(function (_Component) {\n  _inherits(FixedNav, _Component);\n\n  function FixedNav() {\n    _classCallCheck(this, FixedNav);\n\n    return _possibleConstructorReturn(this, (FixedNav.__proto__ || Object.getPrototypeOf(FixedNav)).apply(this, arguments));\n  }\n\n  _createClass(FixedNav, [{\n    key: 'componentWillMount',\n    value: function componentWillMount() {\n      $('.modal').modal();\n    }\n  }, {\n    key: 'showModal',\n    value: function showModal() {\n      $('#search__modal').modal('open');\n    }\n  }, {\n    key: 'showAuthenticationLinks',\n    value: function showAuthenticationLinks() {\n      // Show Sign-in and Sign-up\n      // links only if user isn't signed in\n      if (!this.props.activeState.isAuthenticated) {\n        return _react3.default.createElement(\n          'div',\n          { style: { float: 'left' } },\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signin' },\n              'Sign In'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signup' },\n              'Sign Up'\n            )\n          )\n        );\n      }\n\n      return _react3.default.createElement(\n        _react3.default.Fragment,\n        null,\n        _react3.default.createElement(\n          _IconMenu2.default,\n          {\n            iconButtonElement: _react3.default.createElement(\n              _IconButton2.default,\n              null,\n              _react3.default.createElement(_moreVert2.default, { color: 'white' })\n            ),\n            anchorOrigin: { horizontal: 'left', vertical: 'top' },\n            targetOrigin: { horizontal: 'left', vertical: 'top' }\n          },\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'Hello ' + this.props.activeState.user.lastName }),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'My Events', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/my-events' }) }),\n          _react3.default.createElement(_Divider2.default, null),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'sign out', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/signout' }) })\n        )\n      );\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      return _react3.default.createElement(\n        'div',\n        null,\n        _react3.default.createElement(\n          'div',\n          { className: 'navbar-fixed' },\n          _react3.default.createElement(\n            'nav',\n            { className: 'gradient__bg' },\n            _react3.default.createElement(\n              'div',\n              { className: 'container' },\n              _react3.default.createElement(\n                'div',\n                { className: 'nav-wrapper' },\n                _react3.default.createElement(\n                  _reactRouterDom.Link,\n                  { to: '/', className: 'brand-logo logo' },\n                  _react3.default.createElement(\n                    'p',\n                    null,\n                    'Boots EM'\n                  )\n                ),\n                _react3.default.createElement(\n                  'a',\n                  { href: '#', 'data-activates': 'slide-out', className: 'button-collapse' },\n                  _react3.default.createElement(\n                    'i',\n                    {\n                      className: 'material-icons' },\n                    'menu'\n                  )\n                ),\n                _react3.default.createElement(\n                  'ul',\n                  { id: 'nav-mobile', className: 'right hide-on-med-and-down' },\n                  _react3.default.createElement(\n                    'li',\n                    null,\n                    _react3.default.createElement(\n                      'a',\n                      { onClick: function onClick() {\n                          return _this2.showModal();\n                        }, className: 'modal-trigger', id: 'search__view' },\n                      _react3.default.createElement(\n                        'i',\n                        { className: 'material-icons' },\n                        'search'\n                      )\n                    )\n                  ),\n                  _react3.default.createElement(\n                    'li',\n                    null,\n                    _react3.default.createElement(\n                      _reactRouterDom.Link,\n                      { to: '/centers' },\n                      'List of centers'\n                    )\n                  ),\n                  this.showAuthenticationLinks()\n                )\n              )\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return FixedNav;\n}(_react2.Component));\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    activeState: state.authReducer\n  };\n};\n\nvar matchDispatchToProps = function matchDispatchToProps(dispatch) {\n  return (0, _redux.bindActionCreators)({ signOutRequest: _authActions.signOutRequest }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, { matchDispatchToProps: matchDispatchToProps })(FixedNav);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/bodyComponents/headNav/fixedNav.jsx\n// module id = 188\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/bodyComponents/headNav/fixedNav.jsx?")},269:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _redux = __webpack_require__(23);\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _reactRouterDom = __webpack_require__(27);\n\nvar _IconMenu = __webpack_require__(116);\n\nvar _IconMenu2 = _interopRequireDefault(_IconMenu);\n\nvar _IconButton = __webpack_require__(54);\n\nvar _IconButton2 = _interopRequireDefault(_IconButton);\n\nvar _moreVert = __webpack_require__(89);\n\nvar _moreVert2 = _interopRequireDefault(_moreVert);\n\nvar _MenuItem = __webpack_require__(55);\n\nvar _MenuItem2 = _interopRequireDefault(_MenuItem);\n\nvar _Divider = __webpack_require__(178);\n\nvar _Divider2 = _interopRequireDefault(_Divider);\n\nvar _authActions = __webpack_require__(56);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  Nav: {\n    displayName: 'Nav'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/nav.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/nav.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar Nav = _wrapComponent('Nav')(function (_Component) {\n  _inherits(Nav, _Component);\n\n  function Nav() {\n    _classCallCheck(this, Nav);\n\n    return _possibleConstructorReturn(this, (Nav.__proto__ || Object.getPrototypeOf(Nav)).apply(this, arguments));\n  }\n\n  _createClass(Nav, [{\n    key: 'showAuthenticationLinks',\n    value: function showAuthenticationLinks() {\n      // Show Sign-in and Sign-up\n      // links only if user isn't signed in\n      if (!this.props.activeState.isAuthenticated) {\n        return _react3.default.createElement(\n          'div',\n          { style: { float: 'left' } },\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signin' },\n              'Sign In'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signup' },\n              'Sign Up'\n            )\n          )\n        );\n      }\n\n      return _react3.default.createElement(\n        _react3.default.Fragment,\n        null,\n        _react3.default.createElement(\n          _IconMenu2.default,\n          {\n            iconButtonElement: _react3.default.createElement(\n              _IconButton2.default,\n              null,\n              _react3.default.createElement(_moreVert2.default, { color: 'white' })\n            ),\n            anchorOrigin: { horizontal: 'left', vertical: 'top' },\n            targetOrigin: { horizontal: 'left', vertical: 'top' }\n          },\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'Hello ' + this.props.activeState.user.lastName }),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'My Events', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/my-events' }) }),\n          _react3.default.createElement(_Divider2.default, null),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'sign out', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/signout' }) })\n        )\n      );\n    }\n  }, {\n    key: 'showModal',\n    value: function showModal() {\n      $('#search__modal').modal('open');\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      return _react3.default.createElement(\n        'div',\n        { className: 'navbar-fixed home__nav' },\n        _react3.default.createElement(\n          'nav',\n          { className: 'gradient__bg' },\n          _react3.default.createElement(\n            'div',\n            { className: 'container' },\n            _react3.default.createElement(\n              'div',\n              { className: 'nav-wrapper' },\n              _react3.default.createElement(\n                _reactRouterDom.Link,\n                { to: '/', className: 'brand-logo logo' },\n                _react3.default.createElement(\n                  'p',\n                  null,\n                  'Boots EM'\n                )\n              ),\n              _react3.default.createElement(\n                'ul',\n                { id: 'nav-mobile', className: 'right hide-on-med-and-down' },\n                _react3.default.createElement(\n                  'li',\n                  null,\n                  _react3.default.createElement(\n                    'a',\n                    { onClick: function onClick() {\n                        return _this2.showModal();\n                      }, className: 'modal-trigger', id: 'search__view' },\n                    _react3.default.createElement(\n                      'i',\n                      { className: 'material-icons' },\n                      'search'\n                    )\n                  )\n                ),\n                _react3.default.createElement(\n                  'li',\n                  null,\n                  _react3.default.createElement(\n                    _reactRouterDom.Link,\n                    { to: '/centers' },\n                    'List of centers'\n                  )\n                ),\n                this.showAuthenticationLinks()\n              )\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return Nav;\n}(_react2.Component));\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    activeState: state.authReducer\n  };\n};\n\nvar matchDispatchToProps = function matchDispatchToProps(dispatch) {\n  return (0, _redux.bindActionCreators)({ signOutRequest: _authActions.signOutRequest }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, { matchDispatchToProps: matchDispatchToProps })(Nav);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/bodyComponents/headNav/nav.jsx\n// module id = 269\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/bodyComponents/headNav/nav.jsx?")},28:function(module,exports,__webpack_require__){"use strict";eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nvar CLOUDINARY_URL = exports.CLOUDINARY_URL = 'https://api.cloudinary.com/v1_1/dcbqn1c10/upload';\n\nvar CLOUDINARY_UPLOAD_PRESET = exports.CLOUDINARY_UPLOAD_PRESET = 'z3nqw1ue';\n\n/**\n* ADD_USER action Constant, To add a new user to the store\n* */\nvar ADD_USER = exports.ADD_USER = 'ADD_USER';\n\n/**\n * SET_USER action Constant,\n * To set the current signed in user details and authentication.\n * * */\nvar SET_USER = exports.SET_USER = 'SET_USER';\n\n/**\n * REMOVE_USER action Constant,\n * To remove a user from the database then from the store .\n * * */\nvar REMOVE_USER = exports.REMOVE_USER = 'REMOVE_USER';\n\n/**\n * ADD_CENTER_SUCCESS action Constant,\n * To add a success request action when a new center is being added to the database\n * * */\nvar ADD_CENTER_SUCCESS = exports.ADD_CENTER_SUCCESS = 'ADD_CENTER_SUCCESS';\n\n/**\n * ADD_CENTER_FAlLURE action Constant,\n * To add a failure request action when a new center is being added to the database\n * * */\nvar ADD_CENTER_FAlLURE = exports.ADD_CENTER_FAlLURE = 'ADD_CENTER_FAlLURE';\n\n/**\n * ADD_CENTER_REQUEST action Constant,\n * To add a request action when a new center is being sent to the database\n * * */\nvar ADD_CENTER_REQUEST = exports.ADD_CENTER_REQUEST = 'ADD_CENTER_REQUEST';\n\n/**\n * FETCH_CENTERS action Constant,\n * To fetch centers from the database.\n * * */\nvar FETCH_CENTERS = exports.FETCH_CENTERS = 'FETCH_CENTERS';\n\n/**\n * FETCH_CENTER_DETAIL action Constant,\n * To fetch a center details from the database.\n * * */\nvar FETCH_CENTER_DETAIL = exports.FETCH_CENTER_DETAIL = 'FETCH_CENTER_DETAIL';\n\n/**\n * REMOVE_CENTER action Constant,\n * To remove a center from the database then the store.\n * * */\nvar REMOVE_CENTER = exports.REMOVE_CENTER = 'REMOVE_CENTER';\n\n/**\n * FETCH_RELATED_CENTERS action Constant,\n * To fetch a centers related centers from the database then add to the store.\n * * */\nvar FETCH_RELATED_CENTERS = exports.FETCH_RELATED_CENTERS = 'FETCH_RELATED_CENTERS';\n\n/**\n * EDIT_CENTER_REQUEST action Constant,\n * To add the center details to the store so\n * the details can be gotten and appended to the nodal form.\n * * */\nvar EDIT_CENTER_REQUEST = exports.EDIT_CENTER_REQUEST = 'EDIT_CENTER_REQUEST';\n\n/**\n * EDIT_CENTER action Constant,\n * TO edit and change the center details.\n * * */\nvar EDIT_CENTER = exports.EDIT_CENTER = 'EDIT_CENTER';\n\n/**\n * EDIT_CENTER_FAILURE action Constant,\n * action when edit center request failed.\n * * */\nvar EDIT_CENTER_FAILURE = exports.EDIT_CENTER_FAILURE = 'EDIT_CENTER_FAILURE';\n\n/**\n * FETCH_EVENTS action constants.\n * TO fetch all centers from the database and add to the store.\n * * */\nvar FETCH_EVENTS = exports.FETCH_EVENTS = 'FETCH_EVENTS';\n\n/**\n * ADD_EVENT action constants.\n * To add a new event to the database and to the store.\n * * */\nvar ADD_EVENT = exports.ADD_EVENT = 'ADD_EVENT';\n\n/**\n * EDIT_EVENT_REQUEST action constants.\n * To add a edit request before editing an existing event in the database and the store.\n * * */\nvar EDIT_EVENT_REQUEST = exports.EDIT_EVENT_REQUEST = 'EDIT_EVENT_REQUEST';\n\n/**\n * EDIT_EVENT action constants.\n * To edit an existing event in the database and the store.\n * * */\nvar EDIT_EVENT = exports.EDIT_EVENT = 'EDIT_EVENT';\n\n/**\n * REMOVE_EVENT action constants.\n * To remove an existing event in the database and the store.\n * * */\nvar REMOVE_EVENT = exports.REMOVE_EVENT = 'REMOVE_EVENT';\n\nvar LOADMORE_EVENT_REQUEST = exports.LOADMORE_EVENT_REQUEST = 'LOADMORE_EVENT_REQUEST';\nvar LOADMORE_EVENT_SUCCESS = exports.LOADMORE_EVENT_SUCCESS = 'LOADMORE_EVENT_SUCCESS';\nvar LOADMORE_EVENT_FAILURE = exports.LOADMORE_EVENT_FAILURE = 'LOADMORE_EVENT_FAILURE';\n\nvar LOADMORE_CENTER_REQUEST = exports.LOADMORE_CENTER_REQUEST = 'LOADMORE_CENTER_REQUEST';\nvar LOADMORE_CENTER_SUCCESS = exports.LOADMORE_CENTER_SUCCESS = 'LOADMORE_CENTER_SUCCESS';\nvar LOADMORE_CENTER_FAILURE = exports.LOADMORE_CENTER_FAILURE = 'LOADMORE_CENTER_FAILURE';\n\nvar SEARCH_CENTER_TITLE = exports.SEARCH_CENTER_TITLE = 'SEARCH_CENTER_TITLE';\nvar SEARCH_CENTER_TITLE_FAILED = exports.SEARCH_CENTER_TITLE_FAILED = 'SEARCH_CENTER_TITLE_FAILED';\n\nvar SEARCH_EVENT_TITLE = exports.SEARCH_EVENT_TITLE = 'SEARCH_EVENT_TITLE';\nvar SEARCH_EVENT_TITLE_FAILED = exports.SEARCH_EVENT_TITLE_FAILED = 'SEARCH_EVENT_TITLE_FAILED';\nvar SESSION_EVENTS = exports.SESSION_EVENTS = 'SESSION_EVENTS';\n\nvar IMAGE_PLACEHOLDER = exports.IMAGE_PLACEHOLDER = 'http://placehold.it/600/92c952';\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/actions/index.js\n// module id = 28\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/actions/index.js?")},637:!1,661:function(module,exports,__webpack_require__){"use strict";eval("/* WEBPACK VAR INJECTION */(function(module) {\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _index = __webpack_require__(13);\n\nvar _index2 = _interopRequireDefault(_index);\n\nvar _index3 = __webpack_require__(14);\n\nvar _index4 = _interopRequireDefault(_index3);\n\nvar _react2 = __webpack_require__(0);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _index5 = __webpack_require__(15);\n\nvar _index6 = _interopRequireDefault(_index5);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _redux = __webpack_require__(23);\n\nvar _reactRedux = __webpack_require__(22);\n\nvar _reactRouterDom = __webpack_require__(27);\n\nvar _IconMenu = __webpack_require__(116);\n\nvar _IconMenu2 = _interopRequireDefault(_IconMenu);\n\nvar _IconButton = __webpack_require__(54);\n\nvar _IconButton2 = _interopRequireDefault(_IconButton);\n\nvar _moreVert = __webpack_require__(89);\n\nvar _moreVert2 = _interopRequireDefault(_moreVert);\n\nvar _MenuItem = __webpack_require__(55);\n\nvar _MenuItem2 = _interopRequireDefault(_MenuItem);\n\nvar _Divider = __webpack_require__(178);\n\nvar _Divider2 = _interopRequireDefault(_Divider);\n\nvar _authActions = __webpack_require__(56);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  HeaderBanner: {\n    displayName: 'HeaderBanner'\n  }\n};\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2 = (0, _index6.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/headbanner.jsx',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2 = (0, _index4.default)({\n  filename: '/Users/gabrielmicah/Desktop/projects/event-manager/client/src/components/bodyComponents/headNav/headbanner.jsx',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _index2.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformHmrLibIndexJs2(_UsersGabrielmicahDesktopProjectsEventManagerNode_modulesReactTransformCatchErrorsLibIndexJs2(Component, id), id);\n  };\n}\n\nvar HeaderBanner = _wrapComponent('HeaderBanner')(function (_Component) {\n  _inherits(HeaderBanner, _Component);\n\n  function HeaderBanner(props) {\n    _classCallCheck(this, HeaderBanner);\n\n    var _this = _possibleConstructorReturn(this, (HeaderBanner.__proto__ || Object.getPrototypeOf(HeaderBanner)).call(this, props));\n\n    _this.state = {\n      value: 1\n    };\n    return _this;\n  }\n\n  _createClass(HeaderBanner, [{\n    key: 'showAuthenticationLinks',\n    value: function showAuthenticationLinks() {\n      // Show Sign-in and Sign-up\n      // links only if user isn't signed in\n      if (!this.props.activeState.isAuthenticated) {\n        return _react3.default.createElement(\n          'div',\n          { style: { float: 'left' } },\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signin' },\n              'Sign In'\n            )\n          ),\n          _react3.default.createElement(\n            'li',\n            null,\n            _react3.default.createElement(\n              _reactRouterDom.Link,\n              { to: '/signup' },\n              'Sign Up'\n            )\n          )\n        );\n      }\n\n      return _react3.default.createElement(\n        _react3.default.Fragment,\n        null,\n        _react3.default.createElement(\n          _IconMenu2.default,\n          {\n            iconButtonElement: _react3.default.createElement(\n              _IconButton2.default,\n              null,\n              _react3.default.createElement(_moreVert2.default, { color: 'white' })\n            ),\n            anchorOrigin: { horizontal: 'left', vertical: 'top' },\n            targetOrigin: { horizontal: 'left', vertical: 'top' }\n          },\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'Hello ' + this.props.activeState.user.lastName }),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'My Events', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/my-events' }) }),\n          _react3.default.createElement(_Divider2.default, null),\n          _react3.default.createElement(_MenuItem2.default, { primaryText: 'sign out', containerElement: _react3.default.createElement(_reactRouterDom.Link, { to: '/signout' }) })\n        )\n      );\n    }\n  }, {\n    key: 'showModal',\n    value: function showModal() {\n      $('#search__modal').modal('open');\n    }\n  }, {\n    key: 'showSignUpActionButton',\n    value: function showSignUpActionButton() {\n      if (!this.props.activeState.isAuthenticated) {\n        return _react3.default.createElement(\n          _reactRouterDom.Link,\n          { to: '/signup', className: 'btn blue lighten-2 waves-effect animated fadeInLeft' },\n          'Join Boots Events Manager'\n        );\n      }\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      return _react3.default.createElement(\n        'div',\n        { className: 'header' },\n        _react3.default.createElement(\n          'div',\n          { className: 'header__overlay' },\n          _react3.default.createElement(\n            'div',\n            { className: 'container' },\n            _react3.default.createElement(\n              'nav',\n              { className: 'wow fadeInDown' },\n              _react3.default.createElement(\n                'div',\n                { className: 'nav-wrapper' },\n                _react3.default.createElement(\n                  _reactRouterDom.Link,\n                  { to: '/', className: 'brand-logo logo' },\n                  _react3.default.createElement(\n                    'p',\n                    null,\n                    'Boots EM'\n                  )\n                ),\n                _react3.default.createElement(\n                  'ul',\n                  { id: 'nav-mobile', className: 'right hide-on-med-and-down' },\n                  _react3.default.createElement(\n                    'li',\n                    null,\n                    _react3.default.createElement(\n                      'a',\n                      { onClick: function onClick() {\n                          return _this2.showModal();\n                        }, className: 'modal-trigger', id: 'search__view' },\n                      _react3.default.createElement(\n                        'i',\n                        { className: 'material-icons' },\n                        'search'\n                      )\n                    )\n                  ),\n                  _react3.default.createElement(\n                    'li',\n                    null,\n                    _react3.default.createElement(\n                      _reactRouterDom.Link,\n                      { to: '/centers' },\n                      'List of centers'\n                    )\n                  ),\n                  this.showAuthenticationLinks()\n                )\n              )\n            ),\n            _react3.default.createElement(\n              'div',\n              { className: 'center-align header__detail' },\n              _react3.default.createElement(\n                'h4',\n                { className: 'wow fadeInLeft' },\n                'Worlds Leading Startup events'\n              ),\n              _react3.default.createElement(\n                'p',\n                { className: 'wow fadeInLeft' },\n                'Attend Events around you and Add Events.'\n              ),\n              this.showSignUpActionButton()\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return HeaderBanner;\n}(_react2.Component));\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    activeState: state.authReducer\n  };\n};\n\nvar matchDispatchToProps = function matchDispatchToProps(dispatch) {\n  return (0, _redux.bindActionCreators)({ signOutRequest: _authActions.signOutRequest }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, matchDispatchToProps)(HeaderBanner);\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(10)(module)))\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/components/bodyComponents/headNav/headbanner.jsx\n// module id = 661\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/components/bodyComponents/headNav/headbanner.jsx?")},72:function(module,exports,__webpack_require__){"use strict";eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.loadMoreEvents = exports.deleteEventRequest = exports.editEventAction = exports.editEventRequestAction = exports.fetchSessionEventRequest = exports.fetchEventRequest = exports.createEventRequest = exports.searchEventsDispatch = undefined;\n\nvar _redux = __webpack_require__(23);\n\nvar _axios = __webpack_require__(41);\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _ = __webpack_require__(28);\n\nvar _setAuthenticationToken = __webpack_require__(122);\n\nvar _setAuthenticationToken2 = _interopRequireDefault(_setAuthenticationToken);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\n/**\n *  @API Route String\n * * */\nvar api = '/api/v1/events';\n\n/**\n *  @Event Dispatch Method\n *  @Returns Object\n * * */\nvar eventsDispatchAction = function eventsDispatchAction(type) {\n  var data = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};\n\n  switch (type) {\n    case 'edit':\n      return {\n        type: _.EDIT_EVENT,\n        payload: data\n      };\n\n    case 'edit_request':\n      return {\n        type: _.EDIT_EVENT_REQUEST,\n        payload: data\n      };\n\n    case 'add':\n      return {\n        type: _.ADD_EVENT,\n        payload: data\n      };\n\n    case 'fetch':\n      return {\n        type: _.FETCH_EVENTS,\n        payload: data\n      };\n\n    case 'delete':\n      return {\n        type: _.REMOVE_EVENT,\n        payload: data\n      };\n\n    default:\n      return data;\n  }\n};\n\nvar searchEventsDispatch = exports.searchEventsDispatch = function searchEventsDispatch(data) {\n  return {\n    type: _.SEARCH_EVENT_TITLE,\n    events: data\n  };\n};\n\nvar createEvent = function createEvent(eventData, imgUrl) {\n  return function (dispatch) {\n    var token = localStorage.getItem('jwtToken') ? localStorage.getItem('jwtToken') : false;\n    (0, _setAuthenticationToken2.default)(token);\n    eventData.img_url = imgUrl;\n    return _axios2.default.post(api, eventData).then(function (_ref) {\n      var data = _ref.data;\n\n      dispatch(eventsDispatchAction('add', data.event));\n      Materialize.toast(data.message, 5000, 'teal');\n      location.reload();\n    }).catch(function (err) {\n      console.log(err);\n      Materialize.toast('An Error Occurred..!!!', 5000, 'red');\n    });\n  };\n};\n\n/**\n *  Initial Create Event Request Action\n *  @Returns Promise\n * * */\nvar createEventRequest = exports.createEventRequest = function createEventRequest(eventData) {\n  return function (dispatch) {\n    if (eventData.img_url.name) {\n      var formData = new FormData();\n      formData.append('file', eventData.img_url);\n      formData.append('upload_preset', _.CLOUDINARY_UPLOAD_PRESET);\n      (0, _setAuthenticationToken2.default)(false);\n      return _axios2.default.post(_.CLOUDINARY_URL, formData).then(function (_ref2) {\n        var data = _ref2.data;\n\n        dispatch(createEvent(eventData, data.url));\n      }).catch(function (err) {\n        Materialize.toast('Error in connection', 5000, 'red');\n        console.log(err);\n      });\n    }\n\n    return dispatch(createEvent(eventData, ''));\n  };\n};\n\n/**\n *  @Fetch Event Action\n *  @Returns Object\n * * */\nvar fetchEventRequest = exports.fetchEventRequest = function fetchEventRequest() {\n  return function (dispatch) {\n    return _axios2.default.get(api).then(function (_ref3) {\n      var data = _ref3.data;\n\n      data.loadingmore = false;\n      data.loadmore = false;\n      dispatch(eventsDispatchAction('fetch', data));\n    });\n  };\n};\n\nvar fetchSessionEventRequest = exports.fetchSessionEventRequest = function fetchSessionEventRequest(userId) {\n  return function (dispatch) {\n    return _axios2.default.get(api + '?sessionEvents=' + userId).then(function (_ref4) {\n      var data = _ref4.data;\n\n      dispatch({\n        type: _.SESSION_EVENTS,\n        payload: data.events\n      });\n    });\n  };\n};\n\n/**\n *  @Edit Event Request Action\n *  @Returns Object\n * * */\nvar editEventRequestAction = exports.editEventRequestAction = function editEventRequestAction(data) {\n  return function (dispatch) {\n    return dispatch(eventsDispatchAction('edit_request', data));\n  };\n};\n\nvar editEvent = function editEvent(eventData, imgUrl) {\n  return function (dispatch) {\n    var token = localStorage.getItem('jwtToken') ? localStorage.getItem('jwtToken') : false;\n    (0, _setAuthenticationToken2.default)(token);\n    eventData.img_url = imgUrl;\n    return _axios2.default.put(api + '/' + eventData.eventId, eventData).then(function (_ref5) {\n      var data = _ref5.data;\n\n      if (data.statusCode === 201) {\n        Materialize.toast(data.message, 5000, 'teal');\n        $('#add_event_modal').modal('close');\n        return dispatch(eventsDispatchAction('edit', data.event));\n      }\n      return data;\n    }).catch(function () {\n      Materialize.toast('Error in connection', 5000, 'red');\n    });\n  };\n};\n\n/**\n *  @Edit Event Action\n *  @Returns Object\n * * */\nvar editEventAction = exports.editEventAction = function editEventAction(eventData) {\n  return function (dispatch) {\n    if (eventData.img_url.name) {\n      var formData = new FormData();\n      formData.append('file', eventData.img_url);\n      formData.append('upload_preset', _.CLOUDINARY_UPLOAD_PRESET);\n      (0, _setAuthenticationToken2.default)(false);\n      return _axios2.default.post(_.CLOUDINARY_URL, formData).then(function (_ref6) {\n        var data = _ref6.data;\n\n        dispatch(editEvent(eventData, data.url));\n      }).catch(function () {\n        Materialize.toast('Error in connection', 5000, 'red');\n      });\n    }\n    return dispatch(editEvent(eventData, eventData.img_url));\n  };\n};\n\n/**\n *  @Delete Event Action\n *  @Returns Object\n * * */\nvar deleteEventRequest = exports.deleteEventRequest = function deleteEventRequest(id) {\n  id = parseInt(id, 10);\n  return function (dispatch) {\n    return _axios2.default.delete(api + '/' + id).then(function (_ref7) {\n      var data = _ref7.data;\n\n      if (data.statusCode === 200) {\n        Materialize.toast(data.message, 5000);\n        return dispatch(eventsDispatchAction('delete', data.event));\n      }\n      Materialize.toast(data.message, 5000, 'red');\n      return data;\n    }).catch(function (err) {\n      Materialize.toast(err.message, 5000, 'red');\n    });\n  };\n};\n\n/**\n *  @Load more Event Request Action\n *  @Returns Object\n * * */\nvar loadMoreEvents = exports.loadMoreEvents = function loadMoreEvents(offset) {\n  return function (dispatch) {\n    dispatch({\n      type: _.LOADMORE_EVENT_REQUEST\n    });\n    return _axios2.default.get(api + '?next=' + offset).then(function (_ref8) {\n      var data = _ref8.data;\n\n      if (data.statusCode === 200) {\n        return dispatch({\n          type: _.LOADMORE_EVENT_SUCCESS,\n          payload: data.events\n        });\n      }\n      return dispatch({\n        type: _.LOADMORE_EVENT_FAILURE\n      });\n    });\n  };\n};\n\n//////////////////\n// WEBPACK FOOTER\n// ./client/src/actions/events-actions/index.js\n// module id = 72\n// module chunks = 0\n\n//# sourceURL=webpack:///./client/src/actions/events-actions/index.js?")}});
