+++ client/public/js/hot/hot-update.json
@@ -1,1 +1,1 @@
-{"h":"d3877c06255cbbd6e710","c":{"0":true}}
+{"h":"4e2d305292b9d1b34273","c":{"0":true}}
