+++ package.json
@@ -4,7 +4,7 @@
   "description": "An Event Manager application to create and manage your own event.",
   "main": "./server/app.js",
   "scripts": {
-    "build": "npm run build:client",
+    "build": "npm run clean && npm run build:server && npm run build:client",
     "clean": "rm -rf build && mkdir build",
     "babel-node": "babel-node --preset=es2015",
     "start": "npm run build:server && node -- ./build/app.js",
@@ -18,8 +18,8 @@
     "test": "npm run create:test && npm run migrate:test && npm run test:server",
     "test:local": "npm run drop:test && npm run migrate:test && npm run test:server",
     "test:client": "jest --coverage",
-    "test:server": "cross-env NODE_ENV=test nyc --reporter=html --reporter=text mocha --timeout 20000 --exit --require babel-core/register server/test/*test.js",
-    "test:dev:server": "cross-env NODE_ENV=test nyc --reporter=html --reporter=text mocha ./server/test/* --exit --require babel-node/register --reporter spec",
+    "test:server": "cross-env NODE_ENV=test nyc --reporter=html --reporter=text mocha --timeout 20000 --exit --require babel-core/register server/test/index.js",
+    "test:dev:server": "cross-env NODE_ENV=test nyc --reporter=html --reporter=text mocha ./server/test/index.js --exit --require babel-node/register --reporter spec",
     "test:dev": "npm test && npm run coverage",
     "webpack": "webpack --watch",
     "migrate": "sequelize db:migrate",
