+++ server/config/config.json
@@ -1,0 +1,21 @@
+{
+  "development": {
+    "username": "postgres",
+    "password": "12345",
+    "database": "boots_test",
+    "host": "127.0.0.1",
+    "port": 5432,
+    "operatorsAliases": false,
+    "dialect": "postgres"
+  },
+  "test": {
+    "username": "postgres",
+    "password": "12345",
+    "database": "boots_test",
+    "dialect": "postgres"
+  },
+  "production": {
+    "use_env_variable": "DATABASE_URL",
+    "Dialect": "postgres"
+  }
+}
