+++ server/test/index.js
@@ -1,0 +1,5 @@
+/* eslint-disable no-unused-vars */
+
+import users from './users.test';
+import centers from './center.test';
+import events from './event.test';
